+++ src/Shipment/Manifest.php
@@ -133,8 +133,12 @@ class Manifest
         foreach ($this->orders as $order) {
             $count++;
             $cell_shipment_number = '<td width="' . $this->column_lengths['shipment_number'] . '">' . $order->getTracking() . '</td>';
-            if ($this->show_barcode) {
-                $cell_shipment_number = '<td width="' . $this->column_lengths['shipment_number'] . '" style="line-height: 50%;"><tcpdf method="write1DBarcode" params="' . $pdf->serializeTCPDFtagParameters($this->getBarcodeParams($order->getTracking())) . '" /></td>';
+            if ($this->show_barcode && defined('K_TCPDF_CALLS_IN_HTML') && K_TCPDF_CALLS_IN_HTML === true) {
+                if (method_exists($pdf, 'serializeTCPDFtagParameters')) {
+                    $cell_shipment_number = '<td width="' . $this->column_lengths['shipment_number'] . '" style="line-height: 50%;"><tcpdf method="write1DBarcode" params="' . $pdf->serializeTCPDFtagParameters($this->getBarcodeParams($order->getTracking())) . '" /></td>';
+                } elseif (method_exists($pdf, 'serializeTCPDFtag')) {
+                    $cell_shipment_number = '<td width="' . $this->column_lengths['shipment_number'] . '" style="line-height: 50%;"><tcpdf data="' . $pdf->serializeTCPDFtag('write1DBarcode', $this->getBarcodeParams($order->getTracking())) . '" /></td>';
+                }
             }
             $order_table .= '<tr>
                 <td width = "' . $this->column_lengths['row_number'] . '" align="right">' . $count . '.</td>