+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 - fixed filtering of locations by type
 - adapted to work with PHP 5.6
+- fixed the barcode show when using TCPDF library version 6.7.4 or newer
 
 ## [1.0.17] - Improvements
 - added error message when receiving a "401 Unauthorized" error